MEM: allow ftab-k equal to min-mem-length (fix the signed/unsigned loop hang)
query_mem_bml hung whenever ftab-k was at least min-mem-length because its backward-extension loops used size_t counters against a bound (init_pos - pos_on_r) that goes negative when the ftab already covers the whole seed: at ftab-k == min-mem-length the bound is -1, and the unsigned comparison treats it as huge, so the loop spins forever and indexes out of bounds. The three affected loops now compare as int32_t, so a negative bound correctly skips the loop. With that fixed, ftab-k == min-mem-length is the ideal setting (the ftab seeds the entire minimum window with zero bidirectional extend steps), so the auto-select cap is raised from min-mem-length-1 to min-mem-length and the explicit-ftab guard now rejects only ftab-k strictly greater than min-mem-length (a longer ftab still over-extends the seed left of the intended MEM start and is incorrect). Validated on a small ftab-8/10/12 index: ftab-10 with min-mem 10 no longer hangs and its MEMs are byte-identical to ftab-8; the auto-select now chooses ftab-k=10 for min-mem 10; ftab-12 with min-mem 10 is rejected; and ecoli100 MEM output is unchanged (byte-identical to the 9bb3f72 baseline).

(cherry picked from commit 568bb047ac7ce81e6d6bf875a23ef7511c96b9b6)

+++ src/query_mem.cpp
@@ -44,7 +44,9 @@ bool MoveStructure::query_mem_bml(MoveQuery& mq, int32_t& pos_on_r, int32_t& min
     if (ftab_skip) {
         // If BML skip is known from ftab, find next left end with only backward extension (extend_left slow without ftab)
         MoveInterval fw_interval = bi_interval.fw_interval;
-        for (size_t j = 0; j <= init_pos - pos_on_r; ++j) {
+        // Signed counter: the bound goes negative when the ftab covers the whole seed,
+        // and a size_t one would wrap and loop forever.
+        for (int32_t j = 0; j <= init_pos - pos_on_r; ++j) {
             if (!backward_search_step(query_seq[init_pos - j], fw_interval)) {
                 pos_on_r = init_pos - j + 1;
                 return false;
@@ -55,8 +57,9 @@ bool MoveStructure::query_mem_bml(MoveQuery& mq, int32_t& pos_on_r, int32_t& min
         throw std::runtime_error("Extended past failed ftab");
     }
     else {
-        // Backward extension to find if left end permits a sufficiently long MEM, should equal pos_on_r if true
-        for (size_t j = 0; j <= init_pos - pos_on_r; ++j) {
+        // Backward extension to find if left end permits a sufficiently long MEM, should equal pos_on_r if true.
+        // Signed counter, same reason as above.
+        for (int32_t j = 0; j <= init_pos - pos_on_r; ++j) {
             if (!extend_left(query_seq[init_pos - j], bi_interval)) {
                 pos_on_r = init_pos - j + 1;
                 return false;
@@ -90,7 +93,8 @@ bool MoveStructure::query_mem_bml(MoveQuery& mq, int32_t& pos_on_r, int32_t& min
         MoveInterval fw_interval = initialize_backward_search(mq, init_pos, match_len);
         ++match_len; // TODO: this is still off by one for initialize_backward_search
         --init_pos; // Move to next character left of ftab k-mer or character left of initial range if ftab fails
-        for (i = 0; i <= init_pos - (pos_on_r + 1); ++i) {
+        // Signed comparison, same reason as above.
+        for (i = 0; static_cast<int32_t>(i) <= init_pos - (pos_on_r + 1); ++i) {
             if (!backward_search_step(query_seq[init_pos - i], fw_interval)) {
                 break;
             }
